Fix status issue in app service module

+++ azure-toolkit-libs/azure-toolkit-appservice-lib/src/main/java/com/microsoft/azure/toolkit/lib/appservice/service/impl/FunctionApp.java
@@ -88,14 +88,14 @@ public class FunctionApp extends FunctionAppBase<com.azure.resourcemanager.appse
     @Override
     @Cacheable(cacheName = "appservice/functionapp/{}/slot/{}", key = "${this.name()}/$slotName")
     public IFunctionAppDeploymentSlot deploymentSlot(String slotName) {
-        return new FunctionAppDeploymentSlot(remote(), slotName);
+        return new FunctionAppDeploymentSlot(this, remote(), slotName);
     }
 
     @Override
     @Cacheable(cacheName = "appservice/functionapp/{}/slots", key = "${this.name()}", condition = "!(force&&force[0])")
     public List<IFunctionAppDeploymentSlot> deploymentSlots(boolean... force) {
         return remote().deploymentSlots().list().stream().parallel()
-                .map(functionSlotBasic -> new FunctionAppDeploymentSlot(remote(), functionSlotBasic))
+                .map(functionSlotBasic -> new FunctionAppDeploymentSlot(this, remote(), functionSlotBasic))
                 .collect(Collectors.toList());
     }
 
@@ -127,8 +127,11 @@ public class FunctionApp extends FunctionAppBase<com.azure.resourcemanager.appse
     @Override
     @AzureOperation(name = "function.delete", params = {"this.entity.getName()"}, type = AzureOperation.Type.SERVICE)
     public void delete() {
-        azureClient.functionApps().deleteById(this.id());
-        refresh();
+        if (this.exists()) {
+            this.status(Status.PENDING);
+            azureClient.functionApps().deleteById(this.id());
+            Azure.az(AzureFunction.class).refresh();
+        }
     }
 
     @Override
@@ -138,6 +141,8 @@ public class FunctionApp extends FunctionAppBase<com.azure.resourcemanager.appse
         } finally {
             try {
                 CacheManager.evictCache("appservice/functionapp/{}", this.id());
+                CacheManager.evictCache("appservice/functionapp/{}/slots", this.name());
+                CacheManager.evictCache("appservice/{}/rg/{}/functionapp/{}", String.format("%s/%s/%s", subscriptionId, resourceGroup, name));
             } catch (Throwable e) {
                 log.warn("failed to evict cache", e);
             }
